requirements.txt
Pillow
reportlab
tkinterdnd2
flask
gunicorn
